refactored project directory and introduced styled components. TODO:     - es lint configurations     - refactor other components to use styled components     - pixel perfection with designs for components     - bottom up approach validation
