+++ package.json
@@ -23,8 +23,10 @@
     "react-native": "0.68.2",
     "react-native-modal": "^13.0.1",
     "react-native-simple-time-picker": "^1.3.11",
+    "react-native-status-bar-height": "^2.6.0",
     "react-native-svg": "12.3.0",
-    "react-native-web": "0.17.7"
+    "react-native-web": "0.17.7",
+    "styled-components": "^5.3.5"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
