+++ src/assets/svgs/addIcon.js
