+++ src/assets/svgs/plusIcon.js
