+++ src/screens/Glossary/Glossary.jsx
@@ -1,0 +1,5 @@
+import { Calendar } from '../../components/Calendar';
+
+export const Glossary = () => {
+	return <Calendar />;
+};
